+++ idg-lib/src/CUDA/common/InstanceCUDA.h
@@ -82,7 +82,7 @@ namespace idg {
 
                     void launch_adder(
                         int nr_subgrids,
-                        int grid_size,
+                        long grid_size,
                         int subgrid_size,
                         cu::DeviceMemory& d_metadata,
                         cu::DeviceMemory& d_subgrid,
@@ -90,7 +90,7 @@ namespace idg {
 
                     void launch_adder_unified(
                         int nr_subgrids,
-                        int grid_size,
+                        long grid_size,
                         int subgrid_size,
                         cu::DeviceMemory& d_metadata,
                         cu::DeviceMemory& d_subgrid,
@@ -98,7 +98,7 @@ namespace idg {
 
                     void launch_splitter(
                         int nr_subgrids,
-                        int grid_size,
+                        long grid_size,
                         int subgrid_size,
                         cu::DeviceMemory& d_metadata,
                         cu::DeviceMemory& d_subgrid,
@@ -106,7 +106,7 @@ namespace idg {
 
                     void launch_splitter_unified(
                         int nr_subgrids,
-                        int grid_size,
+                        long grid_size,
                         int subgrid_size,
                         cu::DeviceMemory& d_metadata,
                         cu::DeviceMemory& d_subgrid,
